End cleanup on rustdoc-js tools
